pyproject.toml

[tool.poetry]
name = "jobseekrbot"
version = "0.1.0"
description = "Job Hunter Bot with AI-powered features"
authors = ["Ofir Hamami <admin@hostme.co.il>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = ">=2.3.3,<3.0.0"
flask-login = "^0.6.3"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.1.1"
openai = "^1.57.2"
sqlalchemy = "^2.0.36"
flask-wtf = "^1.2.2"
psycopg2-binary = "^2.9.10"
email-validator = "^2.2.0"
geopy = "^2.4.1"
nest-asyncio = "^1.6.0"
hypercorn = "^0.17.3"
asyncpg = "^0.30.0"
aiohttp = "^3.11.10"
sqlalchemy-utils = "^0.41.2"
python3-saml = "^1.16.0"
requests-oauthlib = "^2.0.0"
certbot = "^3.0.1"
flask-migrate = ">=4.0.7,<5.0.0"
certbot-dns-cloudflare = "^3.0.1"
python-telegram-bot = ">=21.9"
schedule = "^1.2.2"
flask-session = "^0.8.0"
flask-socketio = "^5.4.1"
httpx = ">=0.27.0"
pypdf2 = "^3.0.1"
psutil = "^6.1.1"
asyncio = "^3.4.3"
flask-limiter = "^3.9.2"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
python-docx = "^1.1.2"
docx2txt = "^0.8"
pymupdf = "^1.25.1"
pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=app --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
